fetch/metadata/generated/header-link.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/header-link.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTTP "Link" header</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, rel, test) {
    const iframe = document.createElement('iframe');

    iframe.setAttribute(
      'src',
      '/fetch/metadata/resources/header-link.py' +
        `?location=${encodeURIComponent(url)}&rel=${rel}`
    );

    document.body.appendChild(iframe);
    test.add_cleanup(() => iframe.remove());

    return new Promise((resolve) => {
        iframe.onload = iframe.onerror = resolve;
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpOrigin'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors rel=icon - Not sent to non-trustworthy same-origin destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpOrigin'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors rel=stylesheet - Not sent to non-trustworthy same-origin destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpSameSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors rel=icon - Not sent to non-trustworthy same-site destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpSameSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors rel=stylesheet - Not sent to non-trustworthy same-site destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpCrossSite'], {mime: 'text/html'}),
        'icon',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors rel=icon - Not sent to non-trustworthy cross-site destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpCrossSite'], {mime: 'text/html'}),
        'stylesheet',
        t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors rel=stylesheet - Not sent to non-trustworthy cross-site destination');
  </script>
  </body>
</html>
